pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.13.2",
    "httpx>=0.28.1",
    "pandas>=2.3.3",
    "python-levenshtein>=0.27.3",
    "python-telegram-bot>=22.5",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "rich>=14.2.0",
    "streamlit>=1.52.2",
    "telegram>=0.0.1",
    "thefuzz>=0.22.1",
    "websockets>=15.0.1",
]
